+++ kernel/src/fs/mod.rs
@@ -676,7 +676,7 @@ pub fn init() {
                     if let Ok(f) = etc.create("os-release", Permissions::default()) {
                         f.write(
                             0,
-                            b"NAME=\"VeridianOS\"\nVERSION=\"0.5.1\"\nID=veridian\nPRETTY_NAME=\"VeridianOS v0.5.1\"\n",
+                            b"NAME=\"VeridianOS\"\nVERSION=\"0.5.2\"\nID=veridian\nPRETTY_NAME=\"VeridianOS v0.5.2\"\n",
                         )
                         .ok();
                     }